requirements.txt
google-generativeai
streamlit
openai